Add files via upload
The project aims to assist a local business owner who is interested in purchasing Maven Movies, a movie rental business. The primary goal of the project is to leverage SQL skills to extract and analyze data from the Maven Movies database in order to answer potential acquirer's questions. The business owner primarily owns a restaurant and bar, and therefore has many questions about the movie rental business in general. The offer from the potential acquirer seems generous and we are entertaining his questions to help him make an informed decision about the acquisition.